+++ packages/identifier/src/index.ts
@@ -1,2 +1,3 @@
 export * from './handle'
 export * from './did'
+export * from './nsid'
